Add tipc for PoseC3D.
