+++ test_tipc/configs/PoseC3D/train_infer_python.txt
@@ -1,0 +1,59 @@
+===========================train_params===========================
+model_name:PoseC3D
+python:python3.7
+gpu_list:0
+Global.use_gpu:null|null
+Global.auto_cast:null
+-o epochs:2
+-o output_dir:null
+-o DATASET.batch_size:null
+-o MODEL.backbone.pretrained:null
+train_model_name:null
+train_infer_video_dir:null
+-o DATASET.train.file_path:'data/posec3d_data/mini_ucf101.pkl' -o DATASET.valid.file_path='data/posec3d_data/mini_ucf101.pkl' -o DATASET.test.file_path='data/posec3d_data/mini_ucf101.pkl' -o DATASET.batch_size=1 -o DATASET.valid_batch_size=1
+##
+trainer:norm_train
+norm_train:main.py --validate -c configs/recognition/posec3d/posec3d.yaml --seed 1234
+pact_train:null
+fpgm_train:null
+distill_train:null
+null:null
+null:null
+##
+===========================eval_params===========================
+eval:main.py --test -c configs/recognition/posec3d/posec3d.yaml
+-w:./test_tipc/output/PoseC3D/norm_train_gpus_0_autocast_null/PoseC3D_epoch_00001.pdparams
+##
+===========================infer_params===========================
+-o:inference/PoseC3D
+-p:null
+norm_export:tools/export_model.py -c configs/recognition/posec3d/posec3d.yaml --save_name inference
+quant_export:null
+fpgm_export:null
+distill_export:null
+export1:null
+export2:null
+inference_dir:null
+infer_model:./data/slowfast_4x16.pdparams
+infer_export:tools/export_model.py -c configs/recognition/posec3d/posec3d.yaml
+infer_quant:False
+inference:tools/predict.py --config configs/recognition/posec3d/posec3d.yaml
+--use_gpu:True|False
+--enable_mkldnn:False
+--cpu_threads:1|6
+--batch_size:1|2
+--use_tensorrt:False
+--precision:fp32|fp16
+--model_file:inference.pdmodel
+--input_file:./data/example_UCF101_skeleton.pkl
+null:null
+--enable_benchmark:True
+--params_file:inference.pdiparams
+===========================train_benchmark_params==========================
+batch_size:8
+fp_items:fp32
+epoch:1
+--profiler_options:batch_range=[10,20];state=GPU;tracer_option=Default;profile_path=model.profile
+flags:FLAGS_conv_workspace_size_limit=800
+===========================infer_benchmark_params==========================
+random_infer_input:[{float32,[3, 4, 256, 256]},{float32,[3, 32, 256, 256]}]
